adds flipAnimation
