+++ lib/util/extensions/extensions.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'dart:math';
-import 'dart:ui';
+
 
 import 'package:bonfire/bonfire.dart';
 import 'package:flutter/widgets.dart' as widget;
@@ -14,69 +14,9 @@ export 'movement_extensions.dart';
 export 'npc/npc_extensions.dart';
 export 'player/player_extensions.dart';
 export 'player/rotation_player_extensions.dart';
+export 'image_extensions.dart';
 
 typedef BoolCallback = bool Function();
-
-extension BonfireImageExtension on Image {
-  SpriteAnimation getAnimation({
-    required Vector2 size,
-    required int amount,
-    Vector2? position,
-    double stepTime = 0.1,
-    bool loop = true,
-  }) {
-    return SpriteAnimation.fromFrameData(
-      this,
-      SpriteAnimationData.sequenced(
-        amount: amount,
-        stepTime: stepTime,
-        textureSize: size,
-        loop: loop,
-        texturePosition: position,
-      ),
-    );
-  }
-
-  Sprite getSprite({
-    Vector2? position,
-    Vector2? size,
-  }) {
-    return Sprite(
-      this,
-      srcPosition: position,
-      srcSize: size,
-    );
-  }
-
-  /// Do merge image. Overlaying the images
-  /// @deprecated Use [ImageComposition]
-  Future<Image> overlap(Image other) {
-    PictureRecorder recorder = PictureRecorder();
-    final paint = Paint();
-    Canvas canvas = Canvas(recorder);
-    final totalWidth = max(width, other.width);
-    final totalHeight = max(height, other.height);
-    canvas.drawImage(this, Offset.zero, paint);
-    canvas.drawImage(other, Offset.zero, paint);
-    return recorder.endRecording().toImage(totalWidth, totalHeight);
-  }
-
-  /// Do merge image list. Overlaying the images
-  Future<Image> overlapList(List<Image> others) {
-    PictureRecorder recorder = PictureRecorder();
-    final paint = Paint();
-    Canvas canvas = Canvas(recorder);
-    int totalWidth = width;
-    int totalHeight = height;
-    canvas.drawImage(this, Offset.zero, paint);
-    for (var i in others) {
-      totalWidth = max(totalWidth, i.width);
-      totalHeight = max(totalHeight, i.height);
-      canvas.drawImage(i, Offset.zero, paint);
-    }
-    return recorder.endRecording().toImage(totalWidth, totalHeight);
-  }
-}
 
 extension OffSetExt on Offset {
   Offset copyWith({double? x, double? y}) {
